+++ packages/delivery/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @claudexor/delivery
 
+## 3.13.0
+
+### Patch Changes
+
+- @claudexor/core@3.13.0
+- @claudexor/policy@3.13.0
+- @claudexor/review@3.13.0
+- @claudexor/schema@3.13.0
+- @claudexor/util@3.13.0
+- @claudexor/workspace@3.13.0
+
 ## 3.12.10
 
 ### Patch Changes