+++ packages/harness-fake/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @claudexor/harness-fake
 
+## 3.13.0
+
+### Patch Changes
+
+- @claudexor/core@3.13.0
+- @claudexor/schema@3.13.0
+- @claudexor/util@3.13.0
+
 ## 3.12.10
 
 ### Patch Changes